server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
azure-core
azure-identity
azure-core-auth
azure-ai-voicelive
websockets